refactor: splits -> split
