+++ contracts/README.md
@@ -1,14 +1,8 @@
 # Locker Labs Contracts
 
-This repository contains the smart contracts for Locker Labs. The contracts include various plugins, such as the Savings Plugin and the Split Plugin, that extend functionality and automate token management.
+This repository contains the smart contracts for Locker Labs. The Split Plugin, that extend functionality and automate token management.
 
 ## Plugins
-
-### Savings Plugin
-Enables users to automatically save a portion of their tokens by splitting transfers to a designated savings address.
-
-- Base Sepolia:
-[0x644Be3a596F082CC36D0bD929ABe855180536ac3](https://sepolia.etherscan.io/address/0x644Be3a596F082CC36D0bD929ABe855180536ac3)
 
 ### Split Plugin
 Allows users to split token transfers among multiple recipients automatically, simplifying fund distribution.
