+++ packages/core/index.ts
@@ -3,4 +3,4 @@ export * from "./accounts/index.js";
 export * from "./bridge/index.js";
 export * from "./bridge/providers/cctp/cctpBridge.js";
 export * from "./bridge/types.js";
-export * from "./splits/index.js";
+export * from "./split/index.js";
