+++ packages/core/package.json
@@ -13,7 +13,7 @@
     "ethereum",
     "erc-4337",
     "erc-6900",
-    "splits",
+    "split",
     "round up savings",
     "yield"
   ],
