+++ packages/core/plugingen.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from "@account-kit/plugingen";
-import { SplitPluginGenConfig } from "./splits/def/splitPluginConfig";
+import { SplitPluginGenConfig } from "./split/def/splitPluginConfig";
 import { baseSepolia } from "viem/chains";
 import * as dotenv from "dotenv";
 
@@ -7,7 +7,7 @@ dotenv.config();
 
 export default defineConfig([
   {
-    outDir: "./splits/utils/",
+    outDir: "./split/utils/",
     chain: baseSepolia,
     rpcUrl: process.env.BASE_SEPOLIA_RPC as string,
     plugins: [SplitPluginGenConfig],
