+++ packages/core/split/def/multiOwnerAbi.ts
