+++ packages/core/split/def/splitPluginConfig.ts
