+++ packages/core/split/utils/helpers.ts
