+++ packages/core/split/utils/multiPlugin.ts
