+++ packages/core/split/utils/splitPlugin.ts
